Create screen that list contacts

+++ app/src/main/java/br/com/havebreak/model/Contact.kt
@@ -2,4 +2,4 @@ package br.com.havebreak.model
 
 import java.io.Serializable
 
-data class Contact(var id:Int, var name:String, var username:String, var password:String): Serializable
+data class Contact(var id:Long, var name:String, var username:String, var password:String): Serializable
